More cleanup between tests to avoid Metamorph errors
It seems that store is not automatically cleaned when calling
app.destroy() as this is just regular property on app object.
Furthermore, there is a bug with connectOutlet that prevents cleaning
views that are set with it. I've added a small hack to save all the
views added as outlets to be able to clean them after running tests.