Merge data coming from pusher instead of replacing everything
Pusher events do not always have the entire data for given object.
That's why we need to merge incoming data instead of replacing
everything. Current implementation of merge is just modified version of
load function, so it will be best to either add it to Ember or make the
implementation less dependent on internals.