Fix hiccups with long log views
When scrolling page with long log page was freezing for a moment on
every DOM update (which means at least for a second as we update
duration every second). This was caused by unnecassery reflows - the
main offender was display: box (flexible layout model). It turns out
that newer version of flexible layout, display: flex, behaves much
better. It works on newest versions of Chrome, Safari and on next
version of Opera. Unfortunately Firefox will support it in 18th version,
so I left rules with -moz prefixes for old flexible page model