Load commit from build:started event payload
Commit fields come in with build:started payload. We should split it
into 2 separate keys in the payload, like:

    {
      build:  { .... },
      commit: { .... }
    }

but I can't change pusher v1 API at the moment, due to backwards
compatibility with the old client. We will be able to remove this manual
extraction as soon as we switch to pusher v2 API.