Change shadow for "Follow logs" button
In general, the "light" on web pages come from directly above, so
shadows should be directly below things. This isn't really a shadow,
though, but rather a way to make it seem like the text is "embossed"
into the surface by giving it a slight sheen on the bottom edge, in
which case the shadow definitely shouldn't be shifted left or right.