Fix specs
After update specs fail when there is an unexpected ajax query, because
the response is not an ajax response.

A long term solution is to remove mockjax from specs, because it makes
them brittle and they usually fail after small changes. A better
approach would be to use fixtures, which are available for ember model