Fix flex CSS
We used flex CSS features from both old and new specification.
Furthermore, we weren't using moz and ms prefixes for a new model, which
may have also made it worse in IE and Firefox.

This commit changes CSS to use new flexbox implementation with all the
prefixes (which should not be needed for current versions, but it won't
hurt to keep them for a while).