Revert "Make sorting more in line with what we will do on a server"
This makes the timeline ordering odd when restarting: The repo will be
sorted to be bottom of the timeline while in :created state, and then
pops back up to the top when :started.

One weird thing is that it only seems to behave this way on com
(staging) and enterprise, but not on org (staging).

This reverts commit ce8237b4e75de8a6136123751abc4ceb7b262966.