Change timeout for pusher fallback to 30s
We have a bug that results in sending unneeded ajax queries on missing
log parts. I'm working on a fix, but it may take a while, because I want
to refactor a few parts of the app instead of putting more hacks on
legacy code. To help with the situation I'm increasing timeout for
pusher fallback.