Put jobs with 'received' status to the running jobs list
The problem with running jobs list was that it didn't show jobs that went to the
worker. The result was a confusing UI - when you look at the running jobs list
you want to know how many concurrent jobs you are using, including jobs that are
still waiting for a VM to boot.