Make favicon work properly when viewing job
Current implementation of controllers makes it hard to implement favicon
changes outside of build/job controllers. The problem is that if a job
is viewed, both controller are populated, which means they will both
send state change events to change favicon. This commit adds a check for
build controller, which allows to disable state change events when we're
viewing a job