Add a test for repos-list component
This commit adds 2 unit tests for repos-list. It uses a mocking
technique that I'm playing with lately, ie: don't test parent controller
with nested components classes. Instead of rendering real components,
just mock them and their templates and ensure that proper values are
passed down. This is a move towards good practice of unit tests, but I'm
not yet sure hot it works in the long run. One of the huge benefits is
that changing child component won't affect parent's controller test.
Another advantage is that it's much easier to test various scenarios for
parent controller in case you're in control of what child controller
will render. A downside, however, is that it missed the integration test,i
which I'd like to address in some way later on.