Fix switching log when redirecting without changing routes
When the log changes, but without a route change (like when you switch
from one job to another job), we need to properly clean up and set up a
new log. The best way I figured out is to do it in didUpdateAttrs hook,
but only when the log actually changes. In such situation the old and
the new log should be past to teardown and setup functions.