Only hide #log element, so it's always in the DOM
When a job is not started, we will show a message that the log can't be
shown. If a pusher message with a state change comes after the first log
pusher message, travis-web-log will error out, because in such a
situation a DOM element wouldn't be available. To make it always work,
this commit changes the behaviour to just hide #log element with CSS
instead of using {{#if}}.