Adding some more "about us" info to /about
I got a DM from a friend interested in the job asking if we had
an "about page" that was more than just the headshots, which he
thought were super cool. I took the language from what we've been
including in jobs ads, as a start. I thought I'd submit it for
tweaking and review by those with more design sense than me ;)