More lookup fixes (#504)
* Add a proper interface to flashes service

Before this commit the only way to add a flash message to the flashes
service was to use `loadFlashes` function, which is not a greate API for
a service. In order to make it more intuitive I added 3 functions:
`error`, `notice` and `warning`, each taking only one argument: a
message that you want to show.

* Remove all references to Travis.flash from app

We are currently beginning the process of extracting globally available logic on Travis
into individual services that can be injected as needed.  This fixes some issues around
errors arising from our (improper) use of the new getOwner API.

* Remove unused Broadcast model

* Add acceptance test to exercise build restart notification flow

* Remove outdated run_qunit.js file
* Add new Mirage routes/factories/models
* Upgrade ember-cli-page-object

Shout out to @backspace for some help with debugging some async
behavior.