Convert useV3API to feature flag
The difference between config.featureFlags.useV3API and this.featureFlags.useV3API
has to do with how the feature flag is being used.

We are forced to use the legacy variant whenever we are outside the context of an injectable
object and rely on the imported config. But otherwise, we can use the service.