Move debugging logging behind feature flag
This removes the console clutter we have in our tests (as well as
production). The development environment would contain the logged
information.

My plan for production was to create a debug button that we could
activate at will to debug an issue, so we only have the logging as
needed.