Add sentry integration
This will only log to Sentry in production-like environments (staging and PRs *will* log).
In development, we will receive a console error that Sentry *would have* been notified.

In parallel with this, I have integrated Sentry to post to a #production-web-errors room in
Slack. Email notification is currently disabled, as I do not wish to spam the entire org whenever
something is undefined.

In the future, we might consider having different channels for non-production environments, but that can be addressed
later.