Remove release from build matrix
The upgrade to 2.5.1 now means that we are running the most current version of Ember (release). While we could technically remove default, it is possible that we will be behind by a minor point release intermittently.

I think it makes more sense to remove `release` from our config and assume that we are running the most stable version.

The big win here is that removing a build job will dramatically improve build times. Yay!