Fix Sentry release versions in production environments
Heroku does not appear to build the app within the context of a Git
repository, meaning that the version SHA's are not currently present.

This is being used as part of unique release key for each deploy, which
leads to exceptions not being mapped to releases. This should now be
resolved.