Generate source maps in all environments
This will make source maps a part of our production builds. While this significantly increases the size
of our deploys, my research suggests this will not have an impact on individual client performance, as the .map files
are only downloaded when:
1) source maps are enabled and
2) the developer console is opened

I am not sure if this is entirely accurate, as Sentry will apparently attempt to download the source files, so perhaps this will incur a penalty, but
only when an exception is raised.