Upgrade ember-cli-mirage to latest beta
This brings some breaking changes. First, querying against the schema
now uses the pluralized version of the resource name instead of the
singular.

Second, where we previously were able to pass the result of a schema
query directly, this now no longer returns data suitable for our v3
helpers. Instead, we have to call the `models` property on the object
that is returned.

See the
[CHANGELOG](https://github.com/samselikoff/ember-cli-mirage/blob/master/CHANGELOG.md#020-beta9) for more info.