Persist repository state when enabling repositories
This solves a few issues for new users enabling their first
repositories. Currently, when a user enables their first repository and
subsequently navigates to the dashboard, they do not see any
repositories. This was causing some unexpected errors that made certain
portions of the page unrenderable and led to divergent behavior between
the initial page visit and a refresh (where the data would, of course,
be accurate).

This is because we were missing a link between the state of the store
for repository models and making requests to the /hooks endpoint which
enable them on the backend. To fix this, we now manually inject a
repository model into the store, as the id between a hook and a repo is
shared.  This ensures that code querying the store will get the proper
result.

I think this situation can definitely be improved in the future, perhaps
by merging our concept of hooks with that of repositories.  Were we, for
instance, simply making PATCH requests when toggling a repository's
state (instead of making an RPC-like call to another endpoint), we could
rely on the default behavior of the store to automatically get the
updated state. Given the severity of the issues that new users would
currently encounter, I think we should ship this in its current state
and make improvements later.