Fix spurious error in cron deletion test
Before this change, the cron deletion test produced this error:
Attempted to handle event `notFound` on … while in state root.deleted.saved

The use of the store to load the record and immediately delete it
seems strange to me, might as well just directly delete it. Again
I’m not excited about making a code change while adding tests, but
I couldn’t figure out how to get the test passing with the previous
and seemingly-idiosyncratic implementation.