Add ember-metrics with Segment configuration
I originally had the call to record metrics in the router
as suggested in the ember-metrics README, but I ran into
the problem discussed in this thread:
https://github.com/poteto/ember-metrics/issues/43

For now, I’ve moved it into the application route as this
is only in the experimental phase.

If we proceed with this, we might as well fold in Google
Analytics calls.