Add helper to test for track page events
In the test environment, the TrackCurrentPage mixin is
reporting /tests/index.html as the page, because it uses
document.location.pathname. I was unable to find an Ember
interface that provides the same information you see in
the URL; you can use router.router.state to get the parsed
parameters, but I found no obvious way to get the entire
path.

Eventually I’d like to get a deepEqual in this helper as in
the identify one, but I’ll have to come back to that.