Subscribe to pusher when toggling repos
If a new repository is fetched when doing sync in the application it
will not be subscribed to pusher, because we subscribe only to the
channels that we get back from the user request on initial auth. This
commit fixes the situation so that we always try to subscribe when
activating a repo (if the repo is already on the list it will get
ignored).