Move custom location into conventional position
As noted here: https://github.com/emberjs/ember-inspector/issues/521#issuecomment-190359087

The way the custom location implementation was installed in the
router was causing the Ember Inspector to be unable to load the
route hierarchy. A string is expected for the location implementation
name, to be loaded from app/locations, according to Ember CLI
conventions. This also lets the location itself load the auth service,
via the getOwner API.