Another attempt to get some updated styles merged in (#614)
* Avoiding linting node_modules/ and bower_components/

* Fix as many errors as possible with all rules before proceeding

* Enable multiple rules without issue

* Ignore camelCasing rules in app/{adapters,serializers}

* Use camelcase throughout project

Except where there is contact with an external service:
* app/adapters
* app/serializers
* app/services/store

* Disable func-names for app/utils

This is only temporary until we figure out how best to handle these classes. They should probably just be converted to ES6 classes

* Remove func-names check for now

The changes are way too pervasive and are basically disallowing the
syntax that must be used when disabling prototype extensions for
computed properties.

* Discontinue use of prototype extensions for computed properties

Just used ember-watson to do this. This preps for ember-computed-decorators (way less manual work!)

* Fix spacing errors introduced by new style rules

* Fix line lengths in app/

* Enable additional style rules

Disable max-len (line length) rule for tests, as this would require a
huge effort with little value. The vast majority of the infractions are
due to us supplying meaningful error messages as the third argument to
an assertion.

* Add additional rules and fixes