Add ember-cli-fastboot and ember-network
Ember network is necessary because we do not have access to $.ajax() within node.
This provides a fetch() that is compatible with the browser API to make it run in both
the browser and node.

We could remove this if we ported all network calls to Ember Data in the future, but I do not
think we are ready for that yet.