Fix broken PR tab when loading page directly
This appears to have stopped working recently, but it does not appear
to be related to any changes within the component itself.

I'm adding a guard here as a stop-gap, though we'll need to track down
later what the data changes were that led to this cropping up in the
first place.