Create pro-version feature flag
This does a few things. First, it allows us to target our subscription-based features
directly within acceptance and component tests without having to dig into Ember internals.

Second, it allows us to avoid directly accessing Ember's ENV. The only
place that is now used is within the config itself (and only as a
wrapper for ENV variables passed in when booting the application).

This also brings parity to how we set the pro and enterprise versions.
Before, the pro version was simply defaulted to false. Now, it reads
from the `TRAVIS_PRO` ENV var.