Use proVersion featureFlag instead of config.pro
This still provides the escape valve of accessing the config setting
directly when needed (i.e. the features service can't be injected).
This is now only done from within utils or special objects like the
location hack.

So instead of the previous config.pro, we now reference it through the
service whenever possible like so:

```
if (this.get('features.proVersion')) {
  // do some pro thing
}
```

If that's *not* possible, we can still do the following:

```
if (config.featureFlags['pro-version']) {
  // do some pro thing
}
```