[BUGFIX]: State of running tab persisted correctly (#650)
* Create pro-version feature flag

This does a few things. First, it allows us to target our subscription-based features
directly within acceptance and component tests without having to dig into Ember internals.

Second, it allows us to avoid directly accessing Ember's ENV. The only
place that is now used is within the config itself (and only as a
wrapper for ENV variables passed in when booting the application).

This also brings parity to how we set the pro and enterprise versions.
Before, the pro version was simply defaulted to false. Now, it reads
from the `TRAVIS_PRO` ENV var.

* Use proVersion featureFlag instead of config.pro

This still provides the escape valve of accessing the config setting
directly when needed (i.e. the features service can't be injected).
This is now only done from within utils or special objects like the
location hack.

So instead of the previous config.pro, we now reference it through the
service whenever possible like so:

```
if (this.get('features.proVersion')) {
  // do some pro thing
}
```

If that's *not* possible, we can still do the following:

```
if (config.featureFlags['pro-version']) {
  // do some pro thing
}
```

* Remove unused code
* Fix overriding passed config in pusher util

Wanted to reference our config/environment uniformly, but this clobbers a passed argument.

* Failing test to demonstrate expected behavior that tab state is persisted
* Persist tab state across transitions
* Kill off crons tab for like the hundredth time
* Working implementation with passing tests