Fix broken search tab behavior (#654)
* Fix broken search tab behavior

* Reload normal repository tab when visiting the home screen

This will only occur if we were previously on the search tab though. We want to keep the feeling of a persistent view
between running/my_repos across routes. This is really only necessary because of the implementation detail that search
is actually a tab.

* Perform task instead of triggering action

* Prefer timeout over later runloop invocation

* Remove duplicate tests

* Extract viewSearch contents to task

This creates another task that will be automatically cancelled by its "parent" when
we continue to type and perform tasks.

If we visit /search/foo directly, it is invoked on its own to perform the search.

* Ensure that clicking on sidebar respository result does not change results