Add waiter for logs to render
This is unfortunate but since log rendering happens outside
Ember’s handling, it seems an okay way to await the truncation
message. Another possibility would be embedding some kind of
flag in the Log class and waiting on it to change, but the
innards of that remain mysterious to me.