Move fetching of feature flags to application route
This resolves a few issues that plagued the previous implementation:

First, we were reliant on the feature flag being set for redirects. Typically, the non-blocking nature of the beforeModel hook
was meaning that the feature flag would resolve to false regardless of the actual AJAX response.  No bueno.

Second, we were actually having to deal with the complexity of the request being fired multiple times. We were using ember-concurrency
to drop subsequent requests, but this was unnecessary complexity. Moving the logic to the ApplicationRoute model hook ensures that it
will only fire a single time and block for the duration.

This also resolves another issue with the async behavior, namely some UI
flickering where we initially assumed we would be displaying the old
dashboard.

This does carry with it some potential drawbacks. The first: a broken `/features` endpoint on the API side would essentially hose the app.
I think we should consider moving the flags to local storage and then detecting within the model hook if we already have the hooks enabled.
If so, we could potentially skip the model hook and go with what we have after a certain amount of time.  Another alternative would be to simply
resolve after a certain timeout, meaning that all feature flags would be disabled. This might be jarring depending on what the user is expecting,
but is still better than the app not functioning at all.

Tradeoffs. Tradeoffs. Tradeoffs.

This also removes some unnecessary feature flags that were initially
added for demo purposes.