[WIP] Begin moving more logic into repositories service
Also, fix bug where the "No repos found" (or sth similar) message is displayed in the sidebar.  This does not look great
but is at least present. Looks like this has been broken for some time (went back to Feb. in the commits and could not get the message to display.

Need to confirm that this is actually desired and not just dead code.