Use V3 for cancel/restart and hide buttons when lacking permissions (#698)
* Remove redundant error messages

* Add postV3 method for Ajax service

This will make it easier when the cancel and restart events
move to V3 without needing the unsightly /v3 path prefix.

* Change job/build restart to use V3

* Update permissions-checking for job/build actions

With the V3 API, restarting/cancelling a job requires pull
permissions and debugging requires push. This has broken
the tests, which I’ll fix now.

* Correct computed property dependency

This is what broke the acceptance tests, as the Mirage
permissions factory already granted the proper permissions.

* Disable ESLint for long computed property lines

Is this scandalous? 🤷

* Update component unit tests to refined permissions

These tests don’t seem ideal as they’re mocking everything
of relevance. I’d prefer this to happen in an acceptance
test, which I’ll examine the feasibility of.

* Convert cancellation to use V3

* Add filter to hide no-permissions debug button

* Add missing computed property

* Replace manual computed not with built-in not