Fix Ember CLI version (#749)
I had a CSS rendering crisis locally because of this; Ember CLI
version 2.9.1 matches this semantic versioning string. The
Ember CLI blueprints don’t use the fuzzy matching because of
how easily incompatibilities can creep in, as in this case.

We should be careful in future Ember CLI updates to follow
the blueprint update suggestion.

Thanks to @mixonic for the Yarn reference that helped me
track this down much faster than my painstaking `npm ls`
diff technique was allowing for.