Disable Pusher in test environment
This introduces a “null object” version of the Pusher that
does nothing with the call to subscribe. That’s the only
method I needed to provide an implementation for. We may
want to increase testing around Pusher, in which case we
could replace this with a mock, but for now I just want
to get this working better offline.