Fix Pusher errors by pre-authorizing job channels
This hits our /auth/pusher endpoint with a single job channel to
authorize it before making the subscription call to Pusher.

There was a race condition in the previous code which caused this issue
to only happen intermittently. Now that the subscribe call is
guaranteed to happen after the auth call has succeeded, we shouldn't see
this error anymore.

I'm not super pleased about the implementation, but wasn't sure of a
better way to get at the socketId (normally simply yielded as an
argument to the `bulk_ajax` transport that Pusher calls internally) than
by setting it on the global `Travis.pusher`.