Add script to fetch Redis URLs from Heroku (#783)
The Heroku Redis URLs for the PR/beta/canary deployments change
periodically, which breaks things until I manually update the Redis
environment variables for the Travis environment. This obtains them
at build time from Heroku instead.

This also moves travis-web jobs onto the Trusty beta, because it
includes the heroku command needed to fetch the Redis URLs. This
has the added benefit of including PhantomJS 2.0.0, which comes
pre-installed in that environment, so we no longer need the
.travis.yml lines related to installing that.